Rearrange utility functions and custom tools
