+++ pbsMCP/requirements.txt
@@ -1,6 +1,7 @@
 # Core dependencies for Agentic Workflow Demo
 openai>=1.0.0
 mcp>=1.0.0
+openmm>=8.0.0
 
 # Globus auth (for Sophia LLM access)
 globus-sdk>=3.0.0
